+++ Frends.Mllp.Receive/Frends.Mllp.Receive/Definitions/Options.cs
@@ -1,6 +1,5 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
 
 namespace Frends.Mllp.Receive.Definitions;