+++ Frends.Mllp.Send/CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## [1.2.0] - 2026-04-13
+
+### Changed
+
+- `Connection.Encoding` is now selectable and can be set custom by user
+
 ## [1.1.0] - 2026-03-12
 
 ### Added
 
-- `Connection.Encoding` parameter to allow configuring the character encoding used when sending HL7 messages (default: `UTF-8`)
+- `Connection.Encoding` parameter to allow configuring the character encoding used when sending HL7 messages (default:
+  `UTF-8`)
 
 ## [1.0.0] - 2026-02-05